clientgen: move fakes around to match upstream
