+++ examples/pkg/kcp/clients/clientset/versioned/typed/example/v1/fake/clustertesttype.go
