+++ examples/pkg/kcp/clients/clientset/versioned/typed/example/v1alpha1/fake/testtype.go
