+++ examples/pkg/kcp/clients/clientset/versioned/typed/example/v1beta1/fake/example_client.go
