+++ examples/pkg/kcp/clients/clientset/versioned/typed/example/v2/fake/example_client.go
