+++ examples/pkg/kcp/clients/clientset/versioned/typed/example/v2/fake/testtype.go
