+++ examples/pkg/kcp/clients/clientset/versioned/typed/example3/v1/fake/clustertesttype.go
