+++ examples/pkg/kcp/clients/clientset/versioned/typed/example3/v1/fake/example3_client.go
