+++ examples/pkg/kcp/clients/clientset/versioned/typed/example3/v1/fake/testtype.go
