+++ examples/pkg/kcp/clients/clientset/versioned/typed/existinginterfaces/v1/fake/testtype.go
