+++ examples/pkg/kcp/clients/clientset/versioned/typed/secondexample/v1/fake/clustertesttype.go
