+++ examples/pkg/kcpexisting/clients/clientset/versioned/fake/clientset.go
@@ -39,20 +39,20 @@ import (
 	existinginterfacesv1 "acme.corp/pkg/generated/clientset/versioned/typed/existinginterfaces/v1"
 	secondexamplev1 "acme.corp/pkg/generated/clientset/versioned/typed/secondexample/v1"
 	kcpclient "acme.corp/pkg/kcpexisting/clients/clientset/versioned"
-	fakeexamplev1 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/fake/typed/example/v1"
-	fakeexamplev1alpha1 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/fake/typed/example/v1alpha1"
-	fakeexamplev1beta1 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/fake/typed/example/v1beta1"
-	fakeexamplev2 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/fake/typed/example/v2"
-	fakeexample3v1 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/fake/typed/example3/v1"
-	fakeexistinginterfacesv1 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/fake/typed/existinginterfaces/v1"
-	fakesecondexamplev1 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/fake/typed/secondexample/v1"
 	kcpexamplev1 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/typed/example/v1"
+	fakeexamplev1 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/typed/example/v1/fake"
 	kcpexamplev1alpha1 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/typed/example/v1alpha1"
+	fakeexamplev1alpha1 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/typed/example/v1alpha1/fake"
 	kcpexamplev1beta1 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/typed/example/v1beta1"
+	fakeexamplev1beta1 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/typed/example/v1beta1/fake"
 	kcpexamplev2 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/typed/example/v2"
+	fakeexamplev2 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/typed/example/v2/fake"
 	kcpexample3v1 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/typed/example3/v1"
+	fakeexample3v1 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/typed/example3/v1/fake"
 	kcpexistinginterfacesv1 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/typed/existinginterfaces/v1"
+	fakeexistinginterfacesv1 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/typed/existinginterfaces/v1/fake"
 	kcpsecondexamplev1 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/typed/secondexample/v1"
+	fakesecondexamplev1 "acme.corp/pkg/kcpexisting/clients/clientset/versioned/typed/secondexample/v1/fake"
 )
 
 // NewSimpleClientset returns a clientset that will respond with the provided objects.
