+++ examples/pkg/kcpexisting/clients/clientset/versioned/typed/example/v2/fake/clustertesttype.go
